# backend/swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Kapital API Documentation",
    "version": "1.0.0",
    "description": "API documentation for Kapital backend services"
  },
  "servers": [
    {
      "url": "http://localhost:4001",
      "description": "Local development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/admin/loans": {
      "get": {
        "summary": "Get all approved and disbursed loan applications",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved loans",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/LoanApplication"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized, token is invalid or missing"
          },
          "403": {
            "description": "Forbidden, user is not an admin"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
    "/api/admin/loans/{id}/repayments": {
      "get": {
        "summary": "Get loan repayments (admin only)",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Loan ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Loan repayments retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Access denied, admin privileges required"
          },
          "404": {
            "description": "Loan not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
    "/api/admin/loans/{id}/transactions": {
      "get": {
        "summary": "Get wallet transactions for a specific loan (admin only)",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The loan ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Wallet transactions for the loan",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Access denied, admin privileges required"
          },
          "404": {
            "description": "Loan not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
    "/api/admin/loans/{id}": {
      "get": {
        "summary": "Get individual loan details (admin only)",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Loan ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Loan details retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Access denied, admin privileges required"
          },
          "404": {
            "description": "Loan not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
    "/api/loans": {
      "get": {
        "summary": "Get user's loans",
        "tags": [
          "Loans"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of user's loans (active, pending discharge, and discharged)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "loans": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "principalAmount": {
                            "type": "number"
                          },
                          "outstandingBalance": {
                            "type": "number"
                          },
                          "interestRate": {
                            "type": "number"
                          },
                          "term": {
                            "type": "integer"
                          },
                          "monthlyPayment": {
                            "type": "number"
                          },
                          "nextPaymentDue": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "status": {
                            "type": "string"
                          },
                          "disbursedAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "application": {
                            "type": "object",
                            "properties": {
                              "product": {
                                "type": "object",
                                "properties": {
                                  "name": {
                                    "type": "string"
                                  },
                                  "code": {
                                    "type": "string"
                                  }
                                }
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/loans/{id}": {
      "get": {
        "summary": "Get specific loan details",
        "tags": [
          "Loans"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Loan ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Loan details"
          },
          "404": {
            "description": "Loan not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/loans/{id}/repayments": {
      "get": {
        "summary": "Get loan repayment history",
        "tags": [
          "Loans"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Loan ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Loan repayment history",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "repayments": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "amount": {
                            "type": "number"
                          },
                          "principalAmount": {
                            "type": "number"
                          },
                          "interestAmount": {
                            "type": "number"
                          },
                          "status": {
                            "type": "string"
                          },
                          "dueDate": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "paidAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Loan not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/loans/{id}/transactions": {
      "get": {
        "summary": "Get loan payment transactions from wallet",
        "tags": [
          "Loans"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Loan ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Loan payment transactions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "walletTransactions": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "amount": {
                            "type": "number"
                          },
                          "type": {
                            "type": "string"
                          },
                          "status": {
                            "type": "string"
                          },
                          "description": {
                            "type": "string"
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "updatedAt": {
                            "type": "string",
                            "format": "date-time"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Loan not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/loans/{id}/late-fees": {
      "get": {
        "summary": "Get late fee information for a loan",
        "tags": [
          "Loans"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Loan ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Late fee information for the loan"
          },
          "404": {
            "description": "Loan not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
  },
...
}